Fit method
This method fits a user and item matrices on our data using FunkSVD. It
also stores a lot of useful attributes to our instance to be used later
and for potential visualization by the user. Finally, it stores as
an attribute a df of articles ranked in terms of their popularity.