Predict method
For articles with no interactions but that are in the content dataset,
we find the 10 most similar articles, and average the number of interactions
of the user with these. If the article is in neither datasets, we cannot
predict and return a message saying so. If the user has no interactions,
we cannot predict a number of interactions and return a message saying so.